Add CoreLoader
